requirements.txt
apify-client==0.5.0
asgiref==3.5.1
async-generator==1.10
attrs==21.4.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
cssselect==1.1.0
decorator==4.4.2
Django==4.0.4
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.7.0
google-images-download==2.8.0
h11==0.13.0
idna==3.3
imageio==2.19.2
imageio-ffmpeg==0.4.7
jieba3k==0.35.1
joblib==1.1.0
lxml==4.8.0
moviepy==1.0.3
newspaper3k==0.2.8
nltk==3.7
numpy==1.22.3
outcome==1.1.0
pandas==1.4.2
Pillow==9.1.0
proglog==0.1.10
progressbar==2.5
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-magic-bin==0.4.14
python-pptx==0.6.21
pytz==2022.1
PyYAML==6.0
readability-lxml==0.8.1
regex==2022.4.24
requests==2.27.1
requests-file==1.5.1
selenium==4.1.5
sgmllib3k==1.0.0
simple-image-download==0.5
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tinysegmenter==0.3
tldextract==3.3.0
tqdm==4.64.0
trio==0.20.0
trio-websocket==0.9.2
tzdata==2022.1
urllib3==1.26.9
wsproto==1.1.0
XlsxWriter==3.0.3
